requirements.txt
dify_plugin>=0.2.0,<0.3.0
clickzetta-connector-python>=0.8.51
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
